# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "StudioFlow API",
        "description": "Software Agency & Project Delivery Platform — API-first backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local dev server"
        }
    ],
    "paths": {
...
        "/tasks/{task}/time-logs": {
            "get": {
                "tags": [
                    "TimeLogs"
                ],
                "summary": "List time logs for a task",
                "operationId": "b042690f80d8a3100aa7a43616df6682",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TimeLogs"
                ],
                "summary": "Log time against a task (assignee only)",
                "operationId": "7fdb18df0ec11b5d433915a99155c02d",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hours",
                                    "log_date"
                                ],
                                "properties": {
                                    "hours": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2.5
                                    },
                                    "log_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Logged"
                    },
                    "403": {
                        "description": "Not the assigned user for this task"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the token from /api/login, without the word 'Bearer'.",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}